test: create base class for mockserver tests
Move the boiler-plate code for mockserver tests to a separate
class, so this can easily be re-used for other tests.
