+++ tests/mockserver_tests/test_basics.py
@@ -12,111 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import unittest
-
 from google.cloud.spanner_admin_database_v1.types import spanner_database_admin
-from google.cloud.spanner_v1.testing.mock_database_admin import DatabaseAdminServicer
-from google.cloud.spanner_v1.testing.mock_spanner import (
-    start_mock_server,
-    SpannerServicer,
-)
-import google.cloud.spanner_v1.types.type as spanner_type
-import google.cloud.spanner_v1.types.result_set as result_set
-from google.api_core.client_options import ClientOptions
-from google.auth.credentials import AnonymousCredentials
 from google.cloud.spanner_v1 import (
-    Client,
-    FixedSizePool,
     BatchCreateSessionsRequest,
     ExecuteSqlRequest,
     GetSessionRequest,
 )
-from google.cloud.spanner_v1.database import Database
-from google.cloud.spanner_v1.instance import Instance
-import grpc
+
+from tests.mockserver_tests.mock_server_test_base import (
+    MockServerTestBase,
+    add_select1_result,
+)
 
 
-class TestBasics(unittest.TestCase):
-    server: grpc.Server = None
-    spanner_service: SpannerServicer = None
-    database_admin_service: DatabaseAdminServicer = None
-    port: int = None
-
-    def __init__(self, *args, **kwargs):
-        super(TestBasics, self).__init__(*args, **kwargs)
-        self._client = None
-        self._instance = None
-        self._database = None
-
-    @classmethod
-    def setUpClass(cls):
-        (
-            TestBasics.server,
-            TestBasics.spanner_service,
-            TestBasics.database_admin_service,
-            TestBasics.port,
-        ) = start_mock_server()
-
-    @classmethod
-    def tearDownClass(cls):
-        if TestBasics.server is not None:
-            TestBasics.server.stop(grace=None)
-            TestBasics.server = None
-
-    def _add_select1_result(self):
-        result = result_set.ResultSet(
-            dict(
-                metadata=result_set.ResultSetMetadata(
-                    dict(
-                        row_type=spanner_type.StructType(
-                            dict(
-                                fields=[
-                                    spanner_type.StructType.Field(
-                                        dict(
-                                            name="c",
-                                            type=spanner_type.Type(
-                                                dict(code=spanner_type.TypeCode.INT64)
-                                            ),
-                                        )
-                                    )
-                                ]
-                            )
-                        )
-                    )
-                ),
-            )
-        )
-        result.rows.extend(["1"])
-        TestBasics.spanner_service.mock_spanner.add_result("select 1", result)
-
-    @property
-    def client(self) -> Client:
-        if self._client is None:
-            self._client = Client(
-                project="test-project",
-                credentials=AnonymousCredentials(),
-                client_options=ClientOptions(
-                    api_endpoint="localhost:" + str(TestBasics.port),
-                ),
-            )
-        return self._client
-
-    @property
-    def instance(self) -> Instance:
-        if self._instance is None:
-            self._instance = self.client.instance("test-instance")
-        return self._instance
-
-    @property
-    def database(self) -> Database:
-        if self._database is None:
-            self._database = self.instance.database(
-                "test-database", pool=FixedSizePool(size=10)
-            )
-        return self._database
-
+class TestBasics(MockServerTestBase):
     def test_select1(self):
-        self._add_select1_result()
+        add_select1_result()
         with self.database.snapshot() as snapshot:
             results = snapshot.execute_sql("select 1")
             result_list = []
